+++ Music/BackgroundGameMusic.gd
@@ -1,0 +1,9 @@
+extends AudioStreamPlayer
+
+
+func fade_in():
+	$AnimationPlayer.play("FadeIn")
+
+
+func fade_out():
+	$AnimationPlayer.play("FadeOut")