+++ Utils/Events.gd
@@ -4,3 +4,5 @@ extends Node
 signal footprint(pos, scale)
 # warning-ignore:unused_signal
 signal score_points(points)
+# warning-ignore:unused_signal
+signal add_multiplier(amount)